Move the basic table component to the shared components folder.
