+++ jujugui/static/gui/src/app/components/deployment-flow/pricing/plan-table/test-plan-table.js
@@ -4,7 +4,7 @@
 const React = require('react');
 const enzyme = require('enzyme');
 
-const BasicTable = require('../../../basic-table/basic-table');
+const BasicTable = require('../../../shared/basic-table/basic-table');
 const DeploymentPlanTable = require('./plan-table');
 
 describe('DeploymentPlanTable', () => {
