+++ jujugui/static/gui/src/app/components/deployment-flow/services/test-services.js
@@ -5,7 +5,7 @@ const React = require('react');
 const enzyme = require('enzyme');
 const shapeup = require('shapeup');
 
-const BasicTable = require('../../basic-table/basic-table');
+const BasicTable = require('../../shared/basic-table/basic-table');
 const BudgetTable = require('../../budget-table/budget-table');
 const DeploymentChangeItem = require('../change-item/change-item');
 const DeploymentServices = require('./services');
