+++ jujugui/static/gui/src/app/components/profile/invoice-list/invoice-list.js
@@ -4,7 +4,7 @@
 const PropTypes = require('prop-types');
 const React = require('react');
 
-const BasicTable = require('../../basic-table/basic-table');
+const BasicTable = require('../../shared/basic-table/basic-table');
 const Spinner = require('../../spinner/spinner');
 
 /**
