+++ jujugui/static/gui/src/app/components/shared/basic-table/_basic-table.scss
